Break methods into different files
